+++ metautil.d.ts
@@ -133,3 +133,5 @@ export function unflatObject(
   sourceObject: object,
   fieldNames: Array<string>,
 ): object;
+
+export function isError(instance: object): boolean;