+++ test/utilities.js
@@ -162,6 +162,27 @@ metatests.case(
   },
 );
 
+class ExampleError {}
+
+class ExtendedError extends Error {}
+
+metatests.case(
+  'Error utilities',
+  { metautil },
+  {
+    'metautil.isError': [
+      [new Error('Simple'), true],
+      [new SyntaxError('Bug is here'), true],
+      [new ExampleError('Example'), true],
+      [new ExtendedError('Extended'), true],
+      [{}, false],
+      [[], false],
+      [null, false],
+      [undefined, false],
+    ],
+  },
+);
+
 metatests.case(
   'Math functions',
   { metautil },