+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased][unreleased]
 
+- Add `isError(instance): boolean`
 - Support POST requests with body and custom headers in `fetch()`
 
 ## [3.5.23][] - 2022-08-12